move font_options, text.?pp to font folder
also fix indentation in previous commit
