+++ src/display.cpp
@@ -21,6 +21,7 @@
 #include "cursor.hpp"
 #include "display.hpp"
 #include "fake_unit_manager.hpp"
+#include "font/text.hpp"
 #include "game_preferences.hpp"
 #include "gettext.hpp"
 #include "halo.hpp"
@@ -38,7 +39,6 @@
 #include "synced_context.hpp"
 #include "team.hpp"
 #include "terrain/builder.hpp"
-#include "text.hpp"
 #include "time_of_day.hpp"
 #include "tooltips.hpp"
 #include "tod_manager.hpp"
