+++ src/font/font_id.hpp
@@ -56,9 +56,9 @@ struct font_id
  */
 struct text_chunk
 {
-	text_chunk(subset_id subset) :
-		subset(subset),
-		text()
+	text_chunk(subset_id subset)
+		: subset(subset)
+		, text()
 	{
 	}
