+++ src/reports.cpp
@@ -19,6 +19,7 @@
 //#include "editor/editor_controller.hpp"
 //#include "editor/palette/terrain_palettes.hpp"
 #include "font/sdl_ttf.hpp"
+#include "font/text.hpp"
 #include "game_preferences.hpp"
 #include "gettext.hpp"
 #include "language.hpp"
@@ -28,7 +29,6 @@
 #include "reports.hpp"
 #include "strftime.hpp"
 #include "team.hpp"
-#include "text.hpp"
 #include "tod_manager.hpp"
 #include "units/unit.hpp"
 #include "units/helper.hpp"
