+++ src/storyscreen/render.cpp
@@ -26,11 +26,11 @@
 #include "storyscreen/render.hpp"
 
 #include "config.hpp"
+#include "font/text.hpp"
 #include "image.hpp"
 #include "language.hpp"
 #include "sdl/rect.hpp"
 #include "sound.hpp"
-#include "text.hpp"
 #include "video.hpp"
 #include "widgets/button.hpp"
